OV-748: Stage 4 refactoring. Removing facades that don't do anything.
